+++ CHANGELOG.md
@@ -34,7 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Choosing language between every available locales with autocomplete
 - Dark Theme switch (light theme not implemented)
 - Settings updates are persisted in DB
-- Message modification (modify dialog + messages put route + mesages update in DB)
+- Message modification (modify dialog, messages put route, update in DB)
+- Message deletion (delete dialog, messages delete route, update in DB)
 
 ### Fixed